added average ratings field in service point and service point review…

+++ models/staging/stg_service_ratings.sql
@@ -1,6 +1,6 @@
 select
     id,
-    location_geopoint as location_id,
+    location_geopoint::int as location_id,
     location_name,
     location_activity_types as activity_types,
     rating::integer,
